Authorization overhaul

+++ app/Http/Controllers/TaskController.php
@@ -6,6 +6,7 @@ use App\Task;
 use Illuminate\Http\Request;
 use Illuminate\Support\Carbon;
 use Illuminate\Support\Facades\Auth;
+use Illuminate\Support\Facades\Gate;
 
 class TaskController extends Controller
 {
@@ -21,6 +22,7 @@ class TaskController extends Controller
 
     public function store(Request $request)
     {
+        Gate::authorize('readwrite');
         $request->validate([
             'trigger_id' => 'required|numeric',
             'due_date' => 'required',
@@ -43,6 +45,7 @@ class TaskController extends Controller
 
     public function update(Request $request, Task $task)
     {
+        Gate::authorize('readwrite');
         $this->validate($request, [
             'due_date' => 'sometimes|filled',
             'cost' => 'nullable|numeric',
@@ -68,6 +71,7 @@ class TaskController extends Controller
 
     public function destroy(Task $task)
     {
+        Gate::authorize('readwrite');
         $task->delete();
 
         return $task;
